Refactor: Move DTOs to model package and update imports in Livro, LivroService, and LivroController
