+++ src/main/java/com/portfolio/livros/model/dto/DadosEditarLivro.java
@@ -1,4 +1,4 @@
-package com.portfolio.livros.dto;
+package com.portfolio.livros.model.dto;
 
 public record DadosEditarLivro(Long id, String titulo, String autor) {
 }
